fix(*): 修复图片压缩时，部分图片无法压缩的 BUG
对命令行执行的超时时间进行了延长 (360S)

+++ utils/file_utils.py
@@ -119,11 +119,12 @@ class FileUtils:
                 log.info("执行: " + str_cmd)
                 # os.system(str_cmd)
                 subp = subprocess.Popen(str_cmd)
-                subp.wait(2)
+                subp.wait(360)
 
-            if os.path.exists(output_path):
-                os.remove(input_path)
-                convert_num += 1
+                if subp.poll() == 0 and os.path.exists(output_path):
+                    os.remove(input_path)
+                    convert_num += 1
+
         return convert_num
 
     def upload_tmp_photo(self):
